bagian deletenya bisa

+++ process/mahasiswa/process_delete.php
@@ -2,11 +2,15 @@
 require_once('../../function/helper.php');
 require_once('../../function/koneksi.php');
 
-$id = $_GET['id'];
-
-mysqli_query($koneksi, "DELETE FROM mahasiswa WHERE id='$id'");
-
-header("location: " . BASE_URL . 'dashboard.php?page=datasiswa&status=success');
+    if($_GET['act']=='delete'){
+        $id = $_GET['id'];
+        
+        mysqli_query($koneksi, "DELETE FROM mahasiswa WHERE id='$id'");
+        
+        header("location: " . BASE_URL . 'dashboard.php?page=datasiswa');
+    } else {
+        echo "gagal";
+    }
 
 ?>
